feat(SetReasonStep): add LeaveAlertDialog
